requirements.txt
google-genai
spacy
spacy-layout
pdfminer.six
docx2txt
nltk
streamlit
pandas